CI: Set the target to AuthenticAMD explicitly
When -DHYPERVISOR_TARGET_ARCH= is not set, the build system
will utilize the architecture that the build is currently running on.
Since Github actions seem to run on Intel machines and we are currently
prioritizing our efforts on AMD, we need to set the build to target AMD
explicitly.

In the future, we will want to also enable Intel build targets.

Signed-off-by: Christopher Pelloux <git@chp.io>